move allsets and update_database to scripts
