+++ package.json
@@ -35,7 +35,7 @@
     "pretest": "npm run lint",
     "test": "mocha --reporter spec --exit \"./{,!(node_modules)/**/}*.spec.js\"",
     "setup-env": "node scripts/setup-test-env",
-    "postinstall": "node scripts/setup-test-env && node scripts/postinstall && webpack --config webpack.prod.js"
+    "postinstall": "node scripts/setup-test-env && node scripts/allsets && webpack --config webpack.prod.js"
   },
   "dependencies": {
     "@babel/cli": "^7.7.7",
