--- scripts/postinstall.js
@@ -1,3 +1,0 @@
-// Download Allsets.json
-const allSets = require("../backend/import/allsets");
-allSets.download();
